separate code for email verification logics
